+++ smo/smo/Device.h
@@ -13,11 +13,11 @@ public:
     bool availability() const;
     bool isWaiting() const;
     int getNum() const;
-    double getNumberOfTaken() const;
+    int getNumberOfTaken() const;
 
 private:
     double lambda_;
-    double numberOfTaken_;
+    int numberOfTaken_;
     double releaseTime_;
     int num_;
     bool isAvailable_;
